03_multi_agents_Agno/requirements.txt
agno
python-dotenv
groq
duckduckgo-search
streamlit
googlesearch-python
yfinance
